+++ src/client/i18n/de/de.jsx
@@ -11,7 +11,11 @@
 export const de = {
     'Missing translation': 'Übersetzung fehlt',
     'An unexpected error occurred! Goto {{link}} and try again.': 'Ein unerwarteter Fehler ist aufgetreten! Gehe zur {{link}} und versuche es erneut.',
+    'Oops!': 'Hoppla!',
+    'Homepage': 'Homepage',
     'Version: {{version}}': 'Version: {{version}}',
     'Change to light mode': 'Zu hellem Modus wechseln',
     'Change to dark mode': 'Zu dunklem Modus wechseln',
+    'Settings': 'Einstellungen',
+    'Refresh': 'Aktualisieren',
 }